Flow for the whole app is done.
Improvements needed.

+++ app/src/main/java/com/example/sneakersapp/ui/theme/RatingBar.kt
@@ -45,7 +45,7 @@ fun RatingBarInput(
 
 @Composable
 fun RatingBar(
-    rating: Double,
+    rating: Int,
     modifier: Modifier = Modifier
 ) {
     Row(modifier = modifier) {
